Add an Instructions button that opens a browser.
1. Break ChoiceView out of MainView for maintainability.
2. Add a button labeled "Instructions" that opens a browser on a website.
(On Mac, this opens Safari separately, on iOS it opens an embedded Safari)
